+++ R/zzz.R
@@ -17,7 +17,10 @@
         packageStartupMessage("Using ", threads - 1, " of ", threads, " cores for parallel computing")
         quanteda_options(threads = threads - 1)
     }
-
+    
+    # set dfm print arguments
+    quanteda_options(print_dfm_max_ndoc = 20L)
+    quanteda_options(print_dfm_max_nfeature = 20L)
 }
 
 .onUnload <- function (libpath) {