+++ tests/testthat/test-quanteda_options.R
@@ -6,7 +6,9 @@ test_that("quanteda_options initialize works correctly", {
     expect_equal(
         quanteda_options(),
         list(threads = getOption("quanteda_threads"),
-             verbose = getOption("quanteda_verbose"))
+             verbose = getOption("quanteda_verbose"),
+             print_dfm_max_ndoc = getOption("quanteda_print_dfm_max_ndoc"),
+             print_dfm_max_nfeature = getOption("quanteda_print_dfm_max_nfeature"))
     )
 })
 
@@ -34,6 +36,18 @@ test_that("quanteda_options works correctly to set options", {
     expect_equal(
         quanteda_options("threads"),
         getOption("quanteda_threads")
+    )
+    
+    quanteda_options(print_dfm_max_ndoc = 50L)
+    expect_equal(
+        quanteda_options("print_dfm_max_ndoc"),
+        getOption("quanteda_print_dfm_max_ndoc")
+    )
+    
+    quanteda_options(print_dfm_max_nfeature = 50L)
+    expect_equal(
+        quanteda_options("print_dfm_max_nfeature"),
+        getOption("quanteda_print_dfm_max_nfeature")
     )
 })